requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
sqlalchemy==2.0.35
pyyaml==6.0.2
casbin==1.37.0
python-multipart==0.0.12
httpx==0.27.0
click==8.1.7
python-jose[cryptography]==3.3.0
jwcrypto==1.5.6
cachetools==5.5.0
psycopg2-binary==2.9.10
